requirements.txt
anthropic==0.49.0
numpy==1.23.2
openai==1.66.3
pandas==1.3.4
pydantic==1.10.12
python-dotenv==1.0.1
scikit_learn==0.24.2
tiktoken==0.4.0
